@@ -1,19 +1,17 @@
 {
 	"name": "radiosa-frontend",
-	"version": "0.9.1",
-	"private": true,
+	"version": "0.9.2",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-node": "^5.2.6",
+		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"sass-embedded": "^1.79.4",
-		"svelte": "^4.2.7",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"svelte": "^5.0.0",
 		"vite": "^5.0.3"
-	},
-	"type": "module"
+	}
 }
