@@ -6,5 +6,5 @@ const enabledStationsArray = process.env.ENABLED_STATIONS.split(", ")
 enabledStations = {};
 enabledStationsArray.forEach((channel) => (enabledStations[channel] = allStations[channel]))
 
-fs.writeFileSync("../sveltekit/src/enabled-stations.json", JSON.stringify(enabledStations, null, 4));
+fs.writeFileSync("/radiosa/sveltekit/enabled-stations.json", JSON.stringify(enabledStations, null, 4));
 console.log(`The following stations are enabled: ${process.env.ENABLED_STATIONS}.`);
