@@ -1,3 +1,3 @@
 import fs from "fs";
 
-export default () => JSON.parse(fs.readFileSync("/radiosa/sveltekit/src/enabled-stations.json", "utf8"))
+export default () => JSON.parse(fs.readFileSync("/radiosa/sveltekit/enabled-stations.json", "utf8"))
