@@ -1,4 +1,4 @@
-import getEnabledStations from '$lib/get-enabled-stations.js';
+import getEnabledStations from "$lib/get-enabled-stations.js";
 
 export function load() {
     return { enabledStations: getEnabledStations() };
